+++ packages/zapp-prettier/src/standalone.ts
@@ -1,0 +1,4 @@
+export * from './interfaces';
+export * from './generators';
+export { PrettierProcessor } from './processors/PrettierStandaloneProcessor';
+export * from './schemas';