+++ packages/next-zapp/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @zappjs/next-zapp
 
+## 0.2.0
+
+### Minor Changes
+
+- adds browser build
+
+## 0.1.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @zappjs/prettier@0.6.0
+
 ## 0.1.9
 
 ### Patch Changes