+++ packages/zapp-cli/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @zappjs/cli
 
+## 2.2.0
+
+### Minor Changes
+
+- adds tree view
+
 ## 2.1.0
 
 ### Minor Changes