+++ packages/zapp-prettier/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @zappjs/prettier
 
+## 0.6.0
+
+### Minor Changes
+
+- adds browser build
+
 ## 0.5.1
 
 ### Patch Changes